+++ ExperPort/Analysis/ExtendedStimulus/es_checksph.m
@@ -67,14 +67,14 @@ for rat_i = 1:numel(all_rats)
     
     if plotit,
         if isempty(fignum_in),
-            figure; fignum = gcf;
+            figure; fignum = double(gcf);
         elseif ~ishandle(fignum_in),
             figure(fignum_in);
-            fignum = gcf;
+            fignum = double(gcf);
         else
             figure(fignum_in);
-            clear gcf;
-            fignum = gcf; 
+            clear double(gcf);
+            fignum = double(gcf); 
         end;
 
         ch = get(fignum, 'Children');
