+++ ExperPort/Analysis/NeuraLynx/psthC.m
@@ -1,8 +1,8 @@
 function h=psthC(ev, ts, pre,post,binsz,cnd, meanflg, krn)
 
 h=figure;
-set(gcf, 'Renderer','opengl');
-set(gcf,'Color','w')
+set(double(gcf), 'Renderer','opengl');
+set(double(gcf),'Color','w')
 hold on
 if nargin<6
     cnd=1;
@@ -52,13 +52,13 @@ if ~meanflg,
         %Sundeep Tuteja, 2010-05-10: I have made a change to this line. The
         %axis handle was specified as [] for some reason, causing error
         %messages. It might be necessary to review this line at a later
-        %stage, but for now, I have replaced [] with gca.
-        shadeplot(x/1000,ymn(ci,:)-yst(ci,:),ymn(ci,:)+yst(ci,:),{clrs{ci},gca,.6});
+        %stage, but for now, I have replaced [] with double(gca).
+        shadeplot(x/1000,ymn(ci,:)-yst(ci,:),ymn(ci,:)+yst(ci,:),{clrs{ci},double(gca),.6});
 
     end
 end;
 hold off
-set(gca,'FontSize',20);
+set(double(gca),'FontSize',20);
 
 xlabel('Time from {\bfREF}  in Seconds')
 ylabel('Firing Rate (Hz \pm Std. Err.)')
